WIP - continue implementing and refactoring V2 of the plugin system
