+++ src/Phan/PluginV2/PluginAwareBaseAnalysisVisitor.php
@@ -1,32 +1,30 @@
 <?php declare(strict_types=1);
 namespace Phan\PluginV2;
 
+use Phan\AST\AnalysisVisitor;
 use Phan\AST\Visitor\Element;
 use Phan\CodeBase;
-use Phan\Language\Context;
 use Phan\Issue;
+use Phan\Language\Context;
 use ast\Node;
 
 /**
- * For plugins which define their own post-order analysis behaviors in the analysis phase.
- * Called on a node after PluginAwarePreAnalysisVisitor implementations.
+ * This augments AnalysisVisitor with public and internal methods.
  *
- * - visit<VisitSuffix>(...) (Override these methods)
- * - emitPluginIssue(...) (Call these methods)
- * - emitPluginIssueShort(...)
- * - Public methods from Phan\AST\AnalysisVisitor
- *
- * NOTE: Subclasses should not implement the visit() method unless they absolutely need to.
- * (E.g. if the body would be empty, or if it could be replaced with a small number of more specific methods such as visitFuncDecl, visitVar, etc.)
- *
- * - Phan is able to figure out which methods a subclass implements, and only call the plugin's visitor for those types,
- *   but only when the plugin's visitor does not override the fallback visit() method.
  */
-abstract class PluginAwareAnalysisVisitor extends PluginAwareBaseAnalysisVisitor {
-    /** @var Node|null - Set after the constructor is called */
-    protected $parent_node;
+abstract class PluginAwareBaseAnalysisVisitor extends AnalysisVisitor {
+    use IssueEmitter;  // defines emitPluginIssue
 
-    // Implementations should omit the constructor or call parent::__construct(CodeBase $code_base, Context $context)
+    /**
+     * This is an empty visit() body.
+     * Don't override this unless you need to, analysis is more efficient if Phan knows it doesn't need to call a plugin on a node type.
+     * @see self::isDefinedInSubclass
+     *
+     * @return void
+     */
+    public function visit(Node $node)
+    {
+    }
 
     /**
      * @param string $issue_type
@@ -71,19 +69,31 @@ abstract class PluginAwareAnalysisVisitor extends PluginAwareBaseAnalysisVisitor
             $issue_type_id
         );
     }
-
     // Internal methods used by ConfigPluginSet are below.
     // They aren't useful for plugins.
 
+
     /**
-     * This is a utility function used by ConfigPluginSet
-     * @return void
+     * @return int[] The list of $node->kind values this plugin is capable of analyzing.
      */
-    public static final function staticInvoke(CodeBase $code_base, Context $context, Node $node, Node $parent_node = null)
+    public static final function getHandledNodeKinds() : array
     {
-        $visitor = new static($code_base, $context);
-        $visitor->parent_node = $parent_node;
-        $visitor($node);
+        $defines_visit = self::isDefinedInSubclass('visit');
+        $kinds = [];
+        foreach (Element::VISIT_LOOKUP_TABLE as $kind => $method_name) {
+            if ($defines_visit || self::isDefinedInSubclass($method_name)) {
+                $kinds[] = $kind;
+            }
+        }
+        return $kinds;
     }
-    // End of internal methods.
+    /**
+     * @return bool true if $method_name is defined by the subclass of PluginAwareAnalysisVisitor, and not by PluginAwareAnalysisVisitor or one of it's parents.
+     */
+    private static function isDefinedInSubclass(string $method_name) : bool
+    {
+        $method = new \ReflectionMethod(static::class, $method_name);
+        return is_subclass_of($method->getDeclaringClass()->name, self::class);
+    }
+    // End of methods for internal use.
 }
